chore: re-organize methods and functions
